Fix fish setup

+++ setup/04-fish.sh
@@ -2,5 +2,7 @@
 
 source "$(dirname "$0")/functions.sh"
 
-pp "fish/dotfiles.fish"
-ln -sfnv "$PWD/fish/dotfiles.fish" "$HOME/.config/fish/conf.d/dotfiles.fish" | pcat
+if [[ -e "$HOME/.config/fish/conf.d"/ ]]; then
+  pp "fish/dotfiles.fish"
+  ln -sfnv "$PWD/fish/dotfiles.fish" "$HOME/.config/fish/conf.d/dotfiles.fish" | pcat
+fi
